関数の処理を整理
