+++ web/src/components/chat/MessageWindow.tsx
@@ -1,20 +1,45 @@
-import { Stack, SxProps, Theme } from "@mui/material";
+import { Stack } from "@mui/material";
 import { Box, Paper, Typography } from "@mui/material";
 import { DMOverview, DMRoom, SendMessage, UserID } from "../../common/types";
 import { MessageInput } from "./MessageInput";
 import { useCurrentUserId } from "../../hooks/useCurrentUser";
 import UserAvatar from "../avatar/avatar";
+import { useState, useEffect } from "react";
+import * as chat from "../../api/chat/chat";
 
 type Prop = {
-  data: DMRoom;
-  send: (to: UserID, m: SendMessage) => void;
   room: DMOverview;
-  sx?: SxProps<Theme>;
 };
 
 export function MessageWindow(props: Prop) {
   const id = useCurrentUserId();
-  const { send, room, data } = props;
+  const { room } = props;
+  const [messages, setMessages] = useState<DMRoom | null>(null);
+
+  const sendDMMessage = async (to: UserID, msg: SendMessage): Promise<void> => {
+    await chat.sendDM(to, msg);
+    if (room) {
+      fetchMessages(room.friendId);
+    }
+  };
+
+  const fetchMessages = async (friendId: UserID) => {
+    const newMessages = await chat.getDM(friendId);
+    setMessages(newMessages);
+  };
+
+  useEffect(() => {
+    if (room) {
+      fetchMessages(room.friendId);
+
+      // Set up interval to fetch messages every 5 seconds
+      const intervalId = setInterval(() => fetchMessages(room.friendId), 5000);
+
+      // Clear interval on component unmount or room change
+      return () => clearInterval(intervalId);
+    }
+  }, [room]);
+
   return (
     <>
       <Stack
@@ -38,36 +63,40 @@ export function MessageWindow(props: Prop) {
           flexDirection: "column",
           height: "90%",
           padding: 2,
-          ...props.sx,
         }}
       >
-        <Box sx={{ flexGrow: 1, overflowY: "auto", padding: 1 }}>
-          {data.messages.map((m) => (
-            <Box
-              key={m.id}
-              sx={{
-                display: "flex",
-                justifyContent: m.creator === id ? "flex-end" : "flex-start",
-                marginBottom: 1,
-              }}
-            >
-              <Paper
+        {messages ? (
+          <Box sx={{ flexGrow: 1, overflowY: "auto", padding: 1 }}>
+            {messages.messages.map((m) => (
+              <Box
+                key={m.id}
                 sx={{
                   display: "flex",
-                  maxWidth: "60%",
-                  padding: 1,
-                  borderRadius: 2,
-                  backgroundColor: m.creator === id ? "#DCF8C6" : "#FFF",
-                  boxShadow: 1,
-                  border: 1,
+                  justifyContent: m.creator === id ? "flex-end" : "flex-start",
+                  marginBottom: 1,
                 }}
               >
-                <Typography>{m.content}</Typography>
-              </Paper>
-            </Box>
-          ))}
-        </Box>
-        <MessageInput send={send} room={room} />
+                <Paper
+                  sx={{
+                    display: "flex",
+                    maxWidth: "60%",
+                    padding: 1,
+                    borderRadius: 2,
+                    backgroundColor: m.creator === id ? "#DCF8C6" : "#FFF",
+                    boxShadow: 1,
+                    border: 1,
+                  }}
+                >
+                  <Typography>{m.content}</Typography>
+                </Paper>
+              </Box>
+            ))}
+          </Box>
+        ) : (
+          <Typography>最初のメッセージを送ってみましょう！</Typography>
+        )}
+
+        <MessageInput send={sendDMMessage} room={room} />
       </Box>
     </>
   );
